remove unused `ajb-reuse-windows` variable
Instead of overriding the default action of how a buffer is opened
with `other-window`, I find it more natural to configure your overall emacs
window handling. For example, I am globally preferring using the same
window when possible by reordering the list of
`display-buffer-fallback-action`:

```
(setq display-buffer-fallback-action
      '((display-buffer--maybe-same-window
         display-buffer-reuse-window
         display-buffer-in-previous-window
         display-buffer-use-some-window
         display-buffer--maybe-pop-up-frame-or-window
         display-buffer-pop-up-frame)))
```

This results in not splitting the window again once already split when
using any function that makes use of `other-window`.