Fix: ajb-sort-function has no effect
The current code used let-binding to override `bs-buffer-sort-function` and then
called `bs--show-with-configuration`. But this binding was immediately reset,
since `bs--show-with-configuration` internally calls `bs-set-configuration`
which sets `bs-buffer-sort-function` back to nil. So the buffer lists were
actually never sorted (at least not according to ajb-sort-function).

This commit fixes the issue by advising `bs-set-configuration`. In order to
maintain previous behavior for exiting users, the default value of
`ajb-sort-function` was changed to nil. `defcustom` declaration was also
improved by providing choice of available sort functions.

Also, the existing usage of (obsolete) `defadvice` was replaced by `advice-add`.